hw1/MultiArr/MultiArr.java
/** Multidimensional array
 *  @author Zoe Plaxco
 */

public class MultiArr {

    /**
    {{"hello","you","world"} ,{"how","are","you"}} prints:
    Rows: 2
    Columns: 3

    {{1,3,4},{1},{5,6,7,8},{7,9}} prints:
    Rows: 4
    Columns: 4
    */
    public static void printRowAndCol(int[][] arr) {
        //TODO: Your code here!
    }

    /**
    @param arr: 2d array
    @return maximal value present anywhere in the 2d array
    */
    public static int maxValue(int[][] arr) {
        int result = -1000;
        for (int x = 0; x < arr.length; x++) {
            for (int y = 0; y < arr[x].length; y++) {
                if (arr[x][y] > result){
                    result = arr[x][y];
                }
            }
        }
        return result;
    }

    /**Return an array where each element is the sum of the
    corresponding row of the 2d array*/
    public static int[] allRowSums(int[][] arr) {
        int[] result = new int[arr.length];
        for (int x = 0; x < arr.length; x++) {
            int sum = 0;
            for (int y = 0; y < arr[x].length; y++) {
                sum += arr[x][y];
            }
            result[x] = sum;
        }
        return result;
    }
}
